refactor: move const `OPTIONS_SYMBOL` from types.ts to constants.ts
