+++ src/constants.ts
@@ -1,1 +1,2 @@
 export const NGRX_PACKAGES = ['@ngrx/store', '@ngrx/effects', '@ngrx/signals', '@ngrx/operators'] as const;
+export const OPTIONS_SYMBOL: unique symbol = Symbol('options');
